Rename example modules to examples
